treewide: Implement import on iPadOS
This allows selecting directories from foreign apps and
using them within QML Creator.

This is done using an Objective-C++ type that allows interfacing
with the OS-native document picker, allowing selection of whole
directories and saving their bookmarks within a new SQLite database.

The bookmarks are resolved on the fly, adding a little overhead
to the whole path parsing, but I guess that's unavoidable for now.